Rename samples files
